Support reading multiple bytes at once on GNU C platforms
In order to do multi-byte reads without invalid unaligned memory
accesses, we rely on a gcc feature to create ourselves integer
types that are aligned to one byte. For determining the endianness
of a key we use __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__, macros
provided by gcc and compatible compilers.

The case labels are generated by shifting the individual values
and combining them with a bitwise-or instruction in the generated
source code. This maintains some degree of readability, as the
characters are still present.

To take care of case sensitivity, we can just combine the integer
read with an appropriate run of 0x20. Well almost. If there are
ambiguous characters in the key, we must combine those with 0x20.

The approach we take here is the following:

What we do here is rebuild the tree: For a given node, determine
the minimum position of an ambigous character in any child key.
Now, split the key of all children at that position, re-insert
the value children, and recursively rebuild.

The end result is a uniform trie: At each level of the trie, all
nodes have keys of equal length.