Allow prefix matching with --prefix
In this mode, a word matches one word in the dictionary if the word
in the dictionary is a prefix of the other. For example, if your
dictionary contains the word 'foo', that matches the value 'foobar'.

This will become more useful once we build string tables or let
the user return custom types like strings. Currently you have to
maintain a separate array of strings and index into them.

Suggested-by: Enrico Zini