Disable 2-byte reads, make multi-byte more configurable
2 byte reads negatively affect performance on older intel
architectures as "unaligned memory read, or, cmp, je" is
apparently slower than doing the same thing twice with
single-byte reads.

Performance evaluation shows a huge increase on backup,
which is a ThinkPad Edge 15 with a Core i5-430M. The
other platforms appear to have about the same results
as before, with the exception of the aarch64 case-sensitive
performance, which seems to have gotten lower. But the
compiler has some problems with generating unneeded
instructions on aarch64, so that's probably it's problem.

                     |       before      |       after
function/arch        |     Trie| Trie (*)|     Trie| Trie (*)
---------------------|---------|---------|---------|---------
jak-x230 (amd64)     |     2.35|     2.25|     2.31|     2.25
backup (amd64)       |     1.59|     0.94|     1.78|     1.70
asachi.d.o (arm64)   |     1.31|     1.06|     1.28|     0.85
asachi.d.o (armhf)   |     0.92|     0.85|     0.84|     0.71
asachi.d.o (armel)   |     0.93|     0.74|     0.93|     0.74
plummer.d.o (ppc64el)|     2.06|     1.93|     2.02|     1.83
eller.d.o (mipsel)   |     0.22|     0.23|     0.22|     0.22

Make this configurable: The --multi-byte option now takes
a string of exponents of 2, describing the number of bytes
that may be read. The default is 320, meaning that 8, 4, and
1 byte reads are allowed.