Render C code to match longest prefix
This removes the hardcoded assumption that a node that
has a value is a leaf, and is a more generic approach
to render the tries. It has no effect at all on the
generated output though, because currently, all nodes
with values are in fact leafs (because we maintain one
trie per word length).