Redefine ambiguous to be much more simple
The previous definition of ambiguous was a bit ad hoc as I just
tweaked it until it worked without thinking properly about it.

A character really is ambigous if the bit at position 5 does not
indicate lowercase. If it does, we can use the 0x20 optimization
in a case-insensitive trie without issues.